Move FlaggedStorage to its own file
